Solved AoC2021 Day 16 Part 2

+++ aoc2021/aoc2021-kotlin/src/test/kotlin/de/havox_design/aoc2021/day16/Day16Test.kt
@@ -4,6 +4,7 @@ import de.havox_design.aoc.utils.kotlin.helpers.tests.shouldBe
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
+import java.math.BigInteger
 import java.util.stream.Stream
 
 class Day16Test {
@@ -15,7 +16,7 @@ class Day16Test {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart2")
-    fun testProcessPart2(filename: String, expectedResult: Int) =
+    fun testProcessPart2(filename: String, expectedResult: BigInteger) =
         PacketDecoder(filename).processPart2().shouldBe(expectedResult)
 
     companion object {
@@ -34,14 +35,14 @@ class Day16Test {
         @JvmStatic
         private fun getDataForTestProcessPart2(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample1.txt", 3),
-                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample2.txt", 54),
-                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample3.txt", 7),
-                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample4.txt", 9),
-                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample5.txt", 1),
-                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample6.txt", 0),
-                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample7.txt", 0),
-                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample8.txt", 1)
+                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample1.txt", BigInteger.valueOf(3)),
+                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample2.txt", BigInteger.valueOf(54)),
+                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample3.txt", BigInteger.valueOf(7)),
+                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample4.txt", BigInteger.valueOf(9)),
+                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample5.txt", BigInteger.valueOf(1)),
+                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample6.txt", BigInteger.valueOf(0)),
+                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample7.txt", BigInteger.valueOf(0)),
+                Arguments.of("de/havox_design/aoc2021/day16/day16part2sample8.txt", BigInteger.valueOf(1))
             )
     }
 }
